split editor code and fix drag indicator margins
